refactored getCheapestItems method

+++ src/main/java/com/teamvoy/order/app/mapper/OrderMapper.java
@@ -3,6 +3,7 @@ package com.teamvoy.order.app.mapper;
 import com.teamvoy.order.app.model.Order;
 import com.teamvoy.order.app.model.dto.OrderRequestDto;
 import com.teamvoy.order.app.model.dto.OrderResponseDto;
+import java.math.BigDecimal;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -19,7 +20,7 @@ public class OrderMapper {
     public OrderResponseDto convertFromOrderToResponseDto(Order order) {
         OrderResponseDto responseDto = new OrderResponseDto();
         responseDto.setId(order.getId());
-        responseDto.setPrice(order.getTotalPrice());
+        responseDto.setPrice(BigDecimal.valueOf(order.getTotalPrice()));
         responseDto.setItemName(order.getItemName());
         responseDto.setQuantity(order.getItemsQuantity());
         return responseDto;
