+++ src/test/java/com/teamvoy/order/app/controller/ItemControllerTest.java
@@ -37,7 +37,7 @@ class ItemControllerTest {
         requestDto.setItemName("apple");
         requestDto.setId(3L);
         requestDto.setQuantity(2);
-        requestDto.setPrice(23.0);
+        requestDto.setPrice(BigDecimal.valueOf(23.0));
         mvc.perform(post("/items")
                 .contentType("application/json")
                 .content(objectMapper.writeValueAsString(requestDto)))
